backend/requirements.txt
fastapi
uvicorn
pandas
pydantic
logfire
python-dotenv
scikit-learn
numpy
SQLAlchemy
psycopg2-binary
psycopg2
catboost
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy